Refs #2509 - pxelinux.0 is pkg'd to a new location, copy it from either location if it exists. Is this the correct place for this?  This should be in diskless tools postinst.

+++ src/DCERouter/mkr_postinst.sh
@@ -40,7 +40,8 @@ RunSQL "update Device JOIN DeviceTemplate ON FK_DeviceTemplate=PK_DeviceTemplate
 /usr/pluto/bin/generateRcScripts.sh
 
 mkdir -p /tftpboot/pxelinux.cfg
-cp /usr/lib/syslinux/pxelinux.0 /tftpboot
+[[ -f /usr/lib/syslinux/pxelinux.0 ]] && cp /usr/lib/syslinux/pxelinux.0 /tftpboot || :
+[[ -f /usr/lib/PXELINUX/pxelinux.0 ]] && cp /usr/lib/PXELINUX/pxelinux.0 /tftpboot || :
 
 if [[ "$TestInstallation" == 1 ]]; then 
 	sed -i 's/#log..= \/var\/log\/mysql\/mysql.log/log = \/var\/log\/mysql\/mysql.log/g' /etc/mysql/my.cnf
